move IProduct interface to other projects
