+++ PhoneShop.Client/Program.cs
@@ -2,7 +2,6 @@ using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using PhoneShop.Client;
 using PhoneShop.Client.Services;
-using PhoneShop.Shared.Interfaces;
 using Syncfusion.Blazor;
 
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
@@ -10,7 +9,7 @@ builder.RootComponents.Add<App>("#app");
 builder.RootComponents.Add<HeadOutlet>("head::after");
 
 builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
-builder.Services.AddScoped<IProduct, ClientServices>();
+builder.Services.AddScoped<IProductService, ClientServices>();
 
 builder.Services.AddSyncfusionBlazor();
