+++ PhoneShop.Client/_Imports.razor
@@ -8,9 +8,9 @@
 @using Microsoft.JSInterop
 @using PhoneShop.Client
 @using PhoneShop.Client.Layout
-@using PhoneShop.Shared.Interfaces
+@using PhoneShop.Client.Services
 @using PhoneShop.Shared.Models
 @using Syncfusion.Blazor.Navigations
 
 @inject IJSRuntime js
-@inject IProduct ClientService
+@inject IProductService ClientService
